Refactor home page to comply with server component rules
